Introducing a new abstraction test class to include tests which are only relevant on fake backends, like artificially closing a cursor.
