+++ h2-backend/src/test/java/de/bwaldvogel/mongo/backend/memory/H2BackendTest.java
@@ -1,10 +1,10 @@
 package de.bwaldvogel.mongo.backend.memory;
 
 import de.bwaldvogel.mongo.MongoBackend;
-import de.bwaldvogel.mongo.backend.AbstractBackendTest;
+import de.bwaldvogel.mongo.backend.AbstractFakeBackendTest;
 import de.bwaldvogel.mongo.backend.h2.H2Backend;
 
-public class H2BackendTest extends AbstractBackendTest {
+public class H2BackendTest extends AbstractFakeBackendTest {
 
     @Override
     protected MongoBackend createBackend() throws Exception {
