+++ memory-backend/src/test/java/de/bwaldvogel/mongo/backend/memory/MemoryBackendTest.java
@@ -1,9 +1,9 @@
 package de.bwaldvogel.mongo.backend.memory;
 
 import de.bwaldvogel.mongo.MongoBackend;
-import de.bwaldvogel.mongo.backend.AbstractBackendTest;
+import de.bwaldvogel.mongo.backend.AbstractFakeBackendTest;
 
-public class MemoryBackendTest extends AbstractBackendTest {
+public class MemoryBackendTest extends AbstractFakeBackendTest {
 
     @Override
     protected MongoBackend createBackend() {
