+++ memory-backend/src/test/java/de/bwaldvogel/mongo/backend/memory/MemoryBackendVersion3_4Test.java
@@ -1,9 +1,9 @@
 package de.bwaldvogel.mongo.backend.memory;
 
 import de.bwaldvogel.mongo.MongoBackend;
-import de.bwaldvogel.mongo.backend.AbstractBackendTest;
+import de.bwaldvogel.mongo.backend.AbstractFakeBackendTest;
 
-public class MemoryBackendVersion3_4Test extends AbstractBackendTest {
+public class MemoryBackendVersion3_4Test extends AbstractFakeBackendTest {
 
     @Override
     protected MongoBackend createBackend() {
